+++ src/index.ts
@@ -1,14 +1,14 @@
-import hotkeys, {HotkeysEvent} from 'hotkeys-js';
+import * as Mousetrap from 'mousetrap';
 import {useCallback, useEffect} from "react";
 
 type CallbackFn = (event: KeyboardEvent, handler: HotkeysEvent) => void;
 
-export function useHotkeys(keys: string, callback: CallbackFn, deps: any[] = []) {
+export function useMousetrap(keys: string, callback: CallbackFn, deps: any[] = []) {
   const memoisedCallback = useCallback(callback, deps);
 
   useEffect(() => {
-    hotkeys(keys, memoisedCallback);
+    Mousetrap.bind(keys, memoisedCallback);
 
-    return () => hotkeys.unbind(keys, memoisedCallback);
+    return () => Mousetrap.unbind(keys, memoisedCallback);
   }, [memoisedCallback]);
 }